fix: default_fields link fields like owner
